version 0.4-3

+++ R/bruto.R
@@ -1,5 +1,5 @@
 bruto <-
-function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
+function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2, 
           maxit.select = 20, maxit.backfit = 20, thresh = 1e-04,
           trace.bruto = FALSE, start.linear = TRUE, fit.object, ...)
 {
@@ -11,10 +11,10 @@ function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
     nq <- d[2]
     n <- d[1]
     xnames <- dimnames(x)[[2]]
-    if (!length(xnames))
+    if (!length(xnames)) 
         xnames <- NULL
     ynames <- dimnames(y)[[2]]
-    if (!length(ynames))
+    if (!length(ynames)) 
         ynames <- NULL
     storage.mode(x) <- "double"
     storage.mode(y) <- "double"
@@ -27,22 +27,22 @@ function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
             a2 <- log(100)/log(2)
             a3 <- log(140)/log(2)
             a4 <- log(200)/log(2)
-            cx <- as.integer(cut(n, c(0, 50, 200, 800, 3200)))
-            if (is.na(cx))
+            cx <- as.numeric(cut(n, c(0, 50, 200, 800, 3200)))
+            if (is.na(cx)) 
                 cx <- 5
-            floor(switch(cx, n, 2^(a1 + ((a2 - a1) * (n - 50))/150),
-                2^(a2 + ((a3 - a2) * (n - 200))/600), 2^(a3 +
-                  ((a4 - a3) * (n - 800))/2400), 200 + (n - 3200)^0.2) +
+            floor(switch(cx, n, 2^(a1 + ((a2 - a1) * (n - 50))/150), 
+                2^(a2 + ((a3 - a2) * (n - 200))/600), 2^(a3 + 
+                  ((a4 - a3) * (n - 800))/2400), 200 + (n - 3200)^0.2) + 
                 6)
         }
         check.range <- apply(x, 2, var)
-        if (any(check.range < .Machine$double.eps))
+        if (any(check.range < .Machine$double.eps)) 
             stop(paste("A column of x is constant;",
                        "do not include an intercept column"))
         nkmax <- nknotl(n) - 4
         coef <- matrix(double(nkmax * np * nq), ncol = nq)
         ybar <- apply(y * w, 2, sum)/sum(w)
-        if (start.linear && (nq * cost > n))
+        if (start.linear && (nq * cost > n)) 
             start.linear <- FALSE
         if (start.linear) {
             start.fit <- polyreg(x, y, w)
@@ -64,14 +64,14 @@ function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
         lambda <- double(nq)
         xrange <- matrix(double(2 * nq), 2, nq)
         df <- double(nq)
-        if (missing(dfmax))
+        if (missing(dfmax)) 
             dfmax <- (2 * nkmax)/3
-        if (length(dfmax) != nq)
+        if (length(dfmax) != nq) 
             dfmax <- rep(dfmax, length = nq)
         if (cost > 0) {
             TD <- (n - sum(df))/cost
             TT <- dfmax > TD
-            if (any(TT))
+            if (any(TT)) 
                 dfmax[TT] <- TD
         }
         storage.mode(dfmax) <- "double"
@@ -82,7 +82,7 @@ function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
         nkmax <- fit.object$nkmax
         dfmax <- fit.object$dfmax
         eta <- fit.object$fitted.values
-        if (is.null(eta))
+        if (is.null(eta)) 
             eta <- predict(fit.object, x)
         nk <- fit.object$nk
         knot <- fit.object$knot
@@ -99,13 +99,13 @@ function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
     maxit <- as.integer(c(maxit.select, maxit.backfit))
     names(df) <- xnames
     names(maxit) <- c("selection", "backfitting")
-    gcv.select <- if (maxit.select)
+    gcv.select <- if (maxit.select) 
         matrix(double(maxit.select * nq), nq, maxit.select)
     else double(1)
-    gcv.backfit <- if (maxit.backfit)
+    gcv.backfit <- if (maxit.backfit) 
         matrix(double(maxit.backfit * nq), nq, maxit.backfit)
     else double(1)
-    df.select <- if (maxit.select)
+    df.select <- if (maxit.select) 
         matrix(double(maxit.select * nq), nq, maxit.select)
     else double(1)
     names(lambda) <- xnames
@@ -113,29 +113,29 @@ function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
         .Fortran("bruto",
                  x,
                  as.integer(n),
-                 as.integer(nq),
+                 as.integer(nq), 
                  y,
                  as.integer(np),
                  w,
                  knot = knot,
-                 nkmax = as.integer(nkmax),
+                 nkmax = as.integer(nkmax), 
                  nk = nk,
                  wp,
                  Match = Match,
                  nef = nef,
-                 dfmax = dfmax,
+                 dfmax = dfmax, 
                  cost = cost,
                  lambda = lambda,
                  df = df,
                  coef = coef,
-                 type = type,
+                 type = type, 
                  xrange = xrange,
                  gcv.select = gcv.select,
-                 gcv.backfit = gcv.backfit,
+                 gcv.backfit = gcv.backfit, 
                  df.select = df.select,
                  maxit = maxit,
                  nit = maxit,
-                 fitted.values = eta,
+                 fitted.values = eta, 
                  residuals = y - eta,
                  as.double(thresh),
                  double((2 * np + 2) * ((n + 1) + 1) + (2 * np + 16) *
@@ -164,7 +164,7 @@ function (x, y, w = rep(1, n), wp = rep(1/np, np), dfmax, cost = 2,
     }
     else TT <- NULL
     fit$gcv.backfit <- TT
-    TT <- factor(fit$type, levels = 1:3, labels = c("excluded",
+    TT <- factor(fit$type, levels = 1:3, labels = c("excluded", 
         "linear", "smooth"))
     names(TT) <- xnames
     fit$type <- TT
